game history for logged in player

+++ Backend/src/main/java/com/revature/services/GameHistoryServiceImpl.java
@@ -38,7 +38,7 @@ public class GameHistoryServiceImpl implements GameHistoryService{
     }
 
     @Override
-    public List<GameHistory> getByPersonId(Long id) {
+    public List<GameHistory> getByPersonId(Integer id) {
         return gameHistoryDAO.getByPersonId(id);
     }
 }
